+++ scripts/classes/DieHardDnd5e.js
@@ -1,11 +1,12 @@
 import {dieHardLog} from "../lib/helpers.js";
 
 import DieHardSystem from "./DieHardSystem.js";
-import DieHardFudgeD20Roll from "./DieHardFudgeD20Roll.js";
+import DISABLED_DieHardFudgeD20Roll from "./DISABLED_DieHardFudgeD20Roll.js";
+import {DieHardSetting} from "./DieHard.js";
 
 export default class DieHardDnd5e extends DieHardSystem{
   constructor() {
-    dieHardLog(false, 'DieHardDnd5e - constructor');
+    dieHardLog(false, 'DieHardDnd5e.constructor');
     super();
 
     libWrapper.register('foundry-die-hard', 'CONFIG.Actor.documentClass.prototype.rollAbilitySave', this.actorRollAbilitySave, 'WRAPPER');
@@ -17,8 +18,8 @@ export default class DieHardDnd5e extends DieHardSystem{
 
     libWrapper.register('foundry-die-hard', 'game.dnd5e.dice.D20Roll.prototype._evaluate', this.d20rollEvaluate, 'WRAPPER');
 
-    // See notes in DieHardFudgeD20Roll
-    CONFIG.Dice.DieHardFudgeD20Roll = DieHardFudgeD20Roll;
+    // See notes in DISABLED_DieHardFudgeD20Roll
+    CONFIG.Dice.DieHardFudgeD20Roll = DISABLED_DieHardFudgeD20Roll;
 
     this.totalRollClassName = ["Roll", "D20Roll"]
     this.fudgeWhatOptions = [
@@ -45,19 +46,10 @@ export default class DieHardDnd5e extends DieHardSystem{
     ]
   }
 
-
-
   hookReady() {
     dieHardLog(false, 'Dnd 5e System Hook - Ready');
   }
 
-  isPromise(p) {
-    if (typeof p === 'object' && typeof p.then === 'function') {
-      return true;
-    }
-
-    return false;
-  }
 
   fudgeD20Roll(result, evaluate_options) {
     let functionLogName = 'DieHardDnd5e.fudgeD20Roll'
@@ -81,7 +73,7 @@ export default class DieHardDnd5e extends DieHardSystem{
         SafetyLoopIndex--;
 
         // ToDo: Can a "clone()" or a "reroll()" be used instead?  https://foundryvtt.com/api/Roll.html#clone
-        const new_roll = new DieHardFudgeD20Roll(
+        const new_roll = new DISABLED_DieHardFudgeD20Roll(
           result.formula,
           result.data, {
             flavor: result.options.flavor,
@@ -146,120 +138,79 @@ export default class DieHardDnd5e extends DieHardSystem{
     let result = wrapped.call(evaluate_options)
     // If a fudge re-roll is allowed
     if (fudge){
-      result.then(function(value) {game.settings.get('foundry-die-hard', 'dieHardSettings').system.fudgeD20Roll(value, evaluate_options)})
+      result.then(function(value) {DieHardSetting('dieHardSettings').system.fudgeD20Roll(value, evaluate_options)})
     }
-
-    /*
-    Original WORKING ???
-    //let result = wrapped(evaluate_options); - ORIGINAL
-    let result = wrapped.bind(this)(evaluate_options);
-    // If a fudge re-roll is allowed
-    if (fudge){
-      result.then((value) => game.settings.get('foundry-die-hard', 'dieHardSettings').system.fudgeD20Roll(value, evaluate_options));
-    }
-     */
 
     return result
   }
 
-  /**
-   * Generic wrapper for all rolls
-   * @param options
-   * @param actorId
-   * @param rollType
-   */
-  wrappedRoll(options, actorId, rollType) {
+   wrappedRoll(options, actorId, rollType) {
     let functionLogName = 'DieHardDnd5e.wrappedRoll'
-    //dieHardLog(false, 'DieHardDnd5e.wrappedRoll - this', this);
-    //dieHardLog(false, 'DieHardDnd5e.wrappedRoll - options', options);
-    //dieHardLog(false, 'DieHardDnd5e.wrappedRoll - actorId', actorId);
     dieHardLog(false, functionLogName + ' - rollType', rollType);
 
-    // Check if actor has an active fudge
-    let actorFudges = game.actors.get(actorId).getFlag('foundry-die-hard', 'fudges');
-    if (! Array.isArray(actorFudges)) {
-      actorFudges = []
-    }
-    dieHardLog(false, functionLogName + ' - actorFudges', actorFudges);
-    let fudgeIndex = actorFudges.findIndex(element => { return ((element.whatId === rollType) && (element.statusActive));});
-    if (fudgeIndex !== -1) {
-      dieHardLog(false, functionLogName + ' - active actor fudge', actorFudges[fudgeIndex]);
-      foundry.utils.mergeObject(options, {data: {fudge: true, fudgeOperator: actorFudges[fudgeIndex].operator, fudgeOperatorValue: actorFudges[fudgeIndex].operatorValue, fudgeHow: actorFudges[fudgeIndex].howFormula }});
-      if (actorFudges[fudgeIndex].statusEndless) {
-        dieHardLog(false, functionLogName + ' - fudge is endless');
-      } else {
-        // Disable the fudge
-        actorFudges[fudgeIndex].statusActive = false
-
-        // Delete the fudge from the actor
-        // let deletedFudge = actorFudges.splice(fudgeIndex, 1)
-        game.actors.get(actorId).setFlag('foundry-die-hard', 'fudges', actorFudges);
-        // Check if still have active fudges;
-        this.refreshActiveFudgesIcon();
-      }
-    }
-
-    // Check if user has an active fudge
-    let userFudges = game.users.current.getFlag('foundry-die-hard', 'fudges');
-    if (! Array.isArray(userFudges)) {
-      userFudges = []
-    }
-    dieHardLog(false, functionLogName + ' - userFudges', userFudges);
-    fudgeIndex = userFudges.findIndex(element => { return ((element.whatId === rollType) && (element.statusActive));});
-    if (fudgeIndex !== -1) {
-      dieHardLog(false, functionLogName + ' - active user fudge', userFudges[fudgeIndex]);
-      foundry.utils.mergeObject(options, {data: {fudge: true, fudgeOperator: userFudges[fudgeIndex].operator, fudgeOperatorValue: userFudges[fudgeIndex].operatorValue, fudgeHow: userFudges[fudgeIndex].howFormula }});
-      if (userFudges[fudgeIndex].statusEndless) {
-        dieHardLog(false, functionLogName + ' - fudge is endless');
-      } else {
-        // Disable the fudge
-        userFudges[fudgeIndex].statusActive = false
-
-        // Delete the fudge from the user
-        // let deletedFudge = userFudges.splice(fudgeIndex, 1)
-        game.users.current.setFlag('foundry-die-hard', 'fudges', userFudges);
-        // Check if still have active fudges;
-        this.refreshActiveFudgesIcon();
+    if (! DieHardSetting('fudgeEnabled') ) {
+      dieHardLog(true, functionLogName + ' - Fudge disabled');
+    } else if (DieHardSetting('dieHardSettings').fudgeConfig.globalDisable) {
+      dieHardLog(true, functionLogName + ' - Fudging Globally disabled');
+    } else {
+      // Check if user has an active fudge
+      let userFudge = this.getUserFudge(rollType)
+      if (userFudge !== null) {
+        dieHardLog(false, functionLogName + ' - active user fudge', userFudge);
+        foundry.utils.mergeObject(options, {
+          data: {
+            fudge: true,
+            fudgeOperator: userFudge.operator,
+            fudgeOperatorValue: userFudge.operatorValue,
+            fudgeHow: userFudge.howFormula
+          }
+        });
+        if (userFudge.statusEndless) {
+          dieHardLog(false, functionLogName + ' - fudge is endless');
+        } else {
+          // Disable the fudge
+          this.disableUserFudge(userFudge.id)
+        }
       }
     }
   }
 
   actorRollSkill(wrapped, skillId, options={}) {
     dieHardLog(false, 'DieHardDnd5e.actorRollSkill', this);
-    if (!game.settings.get('foundry-die-hard', 'dieHardSettings').fudgeConfig.globalDisable) {
-      game.settings.get('foundry-die-hard', 'dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollSkill')
+    if (!DieHardSetting('dieHardSettings').fudgeConfig.globalDisable) {
+      DieHardSetting('dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollSkill')
     }
     wrapped(skillId, options);
   }
 
   actorRollAbilitySave(wrapped, abilityId, options={}) {
     dieHardLog(false, 'DieHardDnd5e.actorRollAbilitySave', this);
-    if (!game.settings.get('foundry-die-hard', 'dieHardSettings').fudgeConfig.globalDisable) {
-      game.settings.get('foundry-die-hard', 'dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollAbilitySave')
+    if (!DieHardSetting('dieHardSettings').fudgeConfig.globalDisable) {
+      DieHardSetting('dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollAbilitySave')
     }
     wrapped(abilityId, options);
   }
 
   actorRollAbilityTest(wrapped, abilityId, options={}) {
     dieHardLog(false, 'DieHardDnd5e.actorRollAbilityTest', this);
-    if (!game.settings.get('foundry-die-hard', 'dieHardSettings').fudgeConfig.globalDisable) {
-      game.settings.get('foundry-die-hard', 'dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollAbilityTest')
+    if (!DieHardSetting('dieHardSettings').fudgeConfig.globalDisable) {
+      DieHardSetting('dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollAbilityTest')
     }
     wrapped(abilityId, options);
   }
 
   actorRollDeathSave(wrapped, options={}) {
     dieHardLog(false, 'DieHardDnd5e.actorRollDeathSave', this);
-    if (!game.settings.get('foundry-die-hard', 'dieHardSettings').fudgeConfig.globalDisable) {
-      game.settings.get('foundry-die-hard', 'dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollDeathSave')
+    if (!DieHardSetting('dieHardSettings').fudgeConfig.globalDisable) {
+      DieHardSetting('dieHardSettings').system.wrappedRoll(options, this.id, 'actorRollDeathSave')
     }
     wrapped(options);
   }
 
   entityRollAttack(wrapped, options={}) {
     dieHardLog(false, 'DieHardDnd5e.entityRollAttack', this);
-    if (!game.settings.get('foundry-die-hard', 'dieHardSettings').fudgeConfig.globalDisable) {
-      game.settings.get('foundry-die-hard', 'dieHardSettings').system.wrappedRoll(options, this.actor.id, 'entityRollAttack')
+    if (!DieHardSetting('dieHardSettings').fudgeConfig.globalDisable) {
+      DieHardSetting('dieHardSettings').system.wrappedRoll(options, this.actor.id, 'entityRollAttack')
     }
     wrapped(options);
   }