+++ scripts/classes/DieHardVersionNotification.js
@@ -5,7 +5,7 @@ import {dieHardLog} from "../lib/helpers.js";
 export default class DieHardVersionNotification {
   static checkVersion() {
     let functionLogName = 'DieHardVersionNotification.checkVersion'
-    let notificationVersion = 3
+    let notificationVersion = 4
     if (game.user.isGM && game.user.getFlag('foundry-die-hard', 'versionNotification') !== notificationVersion) {
       dieHardLog(false, functionLogName + ' - Send version notification');
       let commonFooter = "<p>To report problems:<ul><li>open a GitHub issue <a href='https://github.com/UranusBytes/foundry-die-hard/issues' target='_blank'>here</a></li><li>send a message on Discord to <a href='https://discordapp.com/users/530108795796455437' target='_blank'>Glutious#7241</a></li></ul></p>"
@@ -31,6 +31,11 @@ export default class DieHardVersionNotification {
           "<li>Fudge icon adjustments #14</li>" +
           "<li>Fudge config adjustments #15</li>" +
           "<li>Fix fudge whispers going to more than GM #16</li>" +
+          "</ul>",
+        4: "<b>Die Hard v0.0.9</b><p>" +
+            "This version includes the following major changes:</p><ul>" +
+          "<li>Simple and Avg Karma</li>" +
+          "<li>Lots of refactoring to cleanup code</li>" +
           "</ul>",
       }
       let finalMessage = ""